android:maxRecents="3" dev site 는 아래의 설명으로 나와 있는데요, 3이면 너무 작지 않나요? 설계 의도를 다시한번 검토해주세요
The maximum number of tasks rooted at this activity in the overview screen. When this number of entries is reached, the system removes the least-recently used instance from the overview screen. Valid values are 1 through 50 (25 on low memory devices); zero is invalid. This must be an integer value, such as 50. The default value is 16